+++ src/Features/Raytracing/Shape.h
@@ -32,7 +32,7 @@ class Shape
 public:
 	struct Material
 	{
-		enum ShaderType : uint16_t
+		enum ShaderType : uint32_t
 		{
 			TruePBR = 0,
 			Lighting = 1,
@@ -68,7 +68,7 @@ public:
 			}
 		}
 
-		enum ShaderFlags : uint16_t
+		enum ShaderFlags : uint32_t
 		{
 			None = 0,
 			kSpecular = 1 << 0,
@@ -124,7 +124,7 @@ public:
 		RE::BSShader::Type shaderType;
 		REX::EnumSet<RE::BSShaderProperty::EShaderPropertyFlag, std::uint64_t> shaderFlags;
 		RE::BSShaderMaterial::Feature Feature;
-		stl::enumeration<PBRShaderFlags, uint16_t> PBRFlags;
+		stl::enumeration<PBRShaderFlags, uint32_t> PBRFlags;
 
 		MaterialData GetData()
 		{
@@ -142,7 +142,7 @@ public:
 				EnvMaskTexture->GetIndex(),
 				GetShaderType(),
 				GetShaderFlags(),
-				static_cast<uint16_t>(Feature),
+				static_cast<uint32_t>(Feature),
 				PBRFlags.underlying());
 		}
 	};
@@ -209,5 +209,5 @@ public:
 	void CalculateVectors(bool calculateNormal);
 
 	// For PBR shader flags we need to copy exactly what TruePBR does
-	static stl::enumeration<PBRShaderFlags, uint16_t> GetPBRShaderFlags(const BSLightingShaderMaterialPBR* pbrMaterial);
+	static stl::enumeration<PBRShaderFlags, uint32_t> GetPBRShaderFlags(const BSLightingShaderMaterialPBR* pbrMaterial);
 };
